Adapted ll Metal backend enough to compile
