+++ src/backend/metal/Cargo.toml
@@ -1,26 +1,32 @@
 [package]
-name = "gfx_device_metal"
-version = "0.3.0"
-description = "Metal backend for gfx-rs"
+name = "gfx_backend_metal"
+version = "0.1.0"
+description = "Metal API backend for gfx-rs"
 homepage = "https://github.com/gfx-rs/gfx"
 repository = "https://github.com/gfx-rs/gfx"
 keywords = ["graphics", "gamedev"]
 license = "Apache-2.0"
 authors = ["The Gfx-rs Developers"]
 readme = "README.md"
-documentation = "https://docs.rs/gfx_device_metal"
+documentation = "https://docs.rs/gfx_backend_metal"
 workspace = "../../.."
 
 [lib]
-name = "gfx_device_metal"
+name = "gfx_backend_metal"
 
 [dependencies]
-log = "0.3"
 gfx_core = { path = "../../core", version = "0.10" }
-cocoa = "0.9"
-libc = "0.2"
-objc = "0.2"
-objc-foundation = "0.1"
-bit-set = "0.4"
+log = "0.3"
+winit = "0.7"
+scopeguard = "0.3"
+
 metal-rs = "0.4"
+
+objc = "0.2"
 block = "0.1"
+cocoa = "0.9"
+core-foundation = "0.3"
+core-graphics = "0.9"
+io-surface = "0.7"
+
+
