--- src/backend/metal/README.md
@@ -1,13 +1,0 @@
-# gfx_device_metal
-
-[Metal](https://developer.apple.com/metal/) backend for gfx.
-
-## Normalized Coordinates
-
-Render | Depth | Texture
--------|-------|--------
-![render_coordinates](../../../info/gl_render_coordinates.png) | ![depth_coordinates](../../../info/dx_depth_coordinates.png) | ![texture_coordinates](../../../info/dx_texture_coordinates.png)
-
-## Mirroring
-
-TODO
