+++ src/backend/metal/src/command.rs
@@ -1,29 +1,94 @@
-#![allow(missing_docs)]
+use {Surface, Backend};
+use ::native;
+use ::conversions::*;
 
-//use cocoa::foundation::NSRange;
+use std::ops::{Deref, DerefMut, Range};
+use std::sync::{Arc, Mutex};
+use std::cell::UnsafeCell;
 
-use core::{self, pso, shade, state, target, texture, command};
-use core::{IndexType, VertexCount};
-use core::{MAX_VERTEX_ATTRIBUTES, MAX_CONSTANT_BUFFERS, MAX_RESOURCE_VIEWS,
-           MAX_SAMPLERS, MAX_COLOR_TARGETS};
-use core::shade::Stage;
-use {Backend, Resources};
-
-use encoder::MetalEncoder;
-use MTL_MAX_BUFFER_BINDINGS;
-use native::{Buffer, Texture, Pipeline, Rtv, Srv, Dsv};
+use core::{self, mapping, memory, target, pso, state, pool, queue, command, image};
+use core::{VertexCount, VertexOffset, InstanceCount, IndexCount, Viewport};
+use core::{RawSubmission};
+use core::buffer::{IndexBufferView};
+use core::image::{ImageLayout};
+use core::command::{ClearColor, ClearValue, BufferImageCopy, BufferCopy};
+use core::command::{RenderPassInlineEncoder, ImageCopy, SubpassContents};
+use core::command::{ImageResolve};
+use core::pool::{CommandPoolCreateFlags};
 
 use metal::*;
+use cocoa::foundation::NSUInteger;
+use block::{Block, ConcreteBlock};
 
-use std::ptr;
-use std::collections::hash_map::{HashMap, Entry};
-
-#[derive(Clone)]
-pub struct SubmitInfo {
-    pub(crate) command_buffer: MTLCommandBuffer
+pub struct QueueFamily {
 }
 
-unsafe impl Send for SubmitInfo { }
+pub struct CommandQueue(Arc<QueueInner>);
+
+struct QueueInner {
+    queue: MTLCommandQueue,
+}
+
+unsafe impl Send for QueueInner {
+}
+
+impl Drop for QueueInner {
+    fn drop(&mut self) {
+        unsafe {
+            self.queue.release();
+        }
+    }
+}
+
+pub struct CommandPool {
+    queue: Arc<QueueInner>,
+    active_buffers: Vec<CommandBuffer>,
+}
+
+unsafe impl Send for CommandPool {
+}
+unsafe impl Sync for CommandPool {
+}
+
+#[derive(Clone)]
+pub struct CommandBuffer(Arc<UnsafeCell<CommandBufferInner>>);
+
+struct CommandBufferInner {
+    command_buffer: MTLCommandBuffer,
+    encoder_state: EncoderState,
+    viewport: Option<MTLViewport>,
+    scissors: Option<MTLScissorRect>,
+    pipeline_state: Option<MTLRenderPipelineState>, // Unretained
+    vertex_buffers: Vec<(MTLBuffer, pso::BufferOffset)>, // Unretained
+    descriptor_sets: Vec<Option<native::DescriptorSet>>,
+}
+
+unsafe impl Send for CommandBuffer {
+}
+
+impl Drop for CommandBufferInner {
+    fn drop(&mut self) {
+        unsafe {
+            self.command_buffer.release();
+
+            match self.encoder_state {
+                EncoderState::None => {},
+                EncoderState::Blit(encoder) => encoder.release(),
+                EncoderState::Render(encoder) => encoder.release(),
+            }
+        }
+    }
+}
+
+enum EncoderState {
+    None,
+    Blit(MTLBlitCommandEncoder),
+    Render(MTLRenderCommandEncoder),
+}
+
+pub struct SubmitInfo {
+    command_buffer: MTLCommandBuffer
+}
 
 impl Drop for SubmitInfo {
     fn drop(&mut self) {
@@ -31,447 +96,532 @@ impl Drop for SubmitInfo {
     }
 }
 
-pub struct RawCommandBuffer {
-    queue: MTLCommandQueue,
-    device: MTLDevice,
-    encoder: MetalEncoder,
-    should_restore: bool,
-    // TODO: Probably use a more specialized structure, definitely drop SipHash
-    rtv_clear: HashMap<Rtv, MTLClearColor>,
-    dsv_clear: HashMap<Dsv, (Option<f32>, Option<u8>)>,
+impl CommandQueue {
+    pub fn new(device: MTLDevice) -> CommandQueue {
+        CommandQueue(Arc::new(QueueInner {
+            queue: device.new_command_queue(),
+        }))
+    }
+
+    pub unsafe fn device(&self) -> MTLDevice {
+        msg_send![self.0.queue.0, device]
+    }
 }
 
-unsafe impl Send for RawCommandBuffer {}
-
-impl core::CommandBuffer<Backend> for RawCommandBuffer {
-    unsafe fn end(&mut self) -> SubmitInfo {
+impl core::RawCommandQueue<Backend> for CommandQueue {
+    unsafe fn submit_raw(&mut self, submission: RawSubmission<Backend>, fence: Option<&native::Fence>) {
         unimplemented!()
-    }
-}
+        /*for submit in submit_infos {
+            // FIXME: wait for semaphores!
 
-impl RawCommandBuffer {
-    pub fn new(device: MTLDevice, queue: MTLCommandQueue) -> Self {
-        RawCommandBuffer {
-            device: device,
-            queue: queue,
-            encoder: MetalEncoder::new(queue.new_command_buffer()),
-            should_restore: false,
-            rtv_clear: HashMap::new(),
-            dsv_clear: HashMap::new(),
-        }
-    }
-
-    pub fn commit(&mut self, drawable: CAMetalDrawable) {
-        if !self.rtv_clear.is_empty() || !self.dsv_clear.is_empty() {
-            // TODO: should we can find a way to clear buffers anyway (e.g. by issuing a no-op render pass)?
-            warn!("There were unprocessed clear operations in this RawCommandBuffer. Metal only allows clearing \
-            during pixel target binding, so you must bind targets to a render pass between requesting a clear \
-            and committing the RawCommandBuffer.");
-            self.rtv_clear.clear();
-            self.dsv_clear.clear();
-        }
-
-        self.encoder.end_encoding();
-        self.encoder.commit_command_buffer(drawable, false);
-        self.encoder.reset();
-
-        self.should_restore = false;
-    }
-
-    fn ensure_render_encoder(&mut self) {
-        if !self.encoder.has_command_buffer() {
-            self.encoder.start_command_buffer(self.queue.new_command_buffer());
-        }
-
-        if !self.encoder.is_render_encoding() {
-            self.encoder.begin_render_encoding();
-        }
-
-        if self.should_restore {
-            self.encoder.restore_render_state();
-            self.should_restore = false;
-        }
-    }
-}
-
-impl command::Buffer<Resources> for RawCommandBuffer {
-    fn reset(&mut self) {
-        self.encoder.reset();
-    }
-
-    fn bind_pipeline_state(&mut self, pso: Pipeline) {
-        self.encoder.set_render_pipeline_state(pso.pipeline);
-        self.encoder.set_front_facing_winding(pso.winding);
-        self.encoder.set_cull_mode(pso.cull);
-        self.encoder.set_triangle_fill_mode(pso.fill);
-
-        // TODO(fkaa): do we need max value?
-        self.encoder.set_depth_bias(pso.depth_bias as f32 ,
-                                    pso.slope_scaled_depth_bias as f32,
-                                    0f32);
-
-        if let Some(depth_state) = pso.depth_stencil {
-            self.encoder.set_depth_stencil_state(depth_state);
-        }
-    }
-
-    fn bind_vertex_buffers(&mut self, vbs: pso::VertexBufferSet<Resources>) {
-        let mut vb_count = 0;
-        for i in 0..MAX_VERTEX_ATTRIBUTES {
-            if let Some((buffer, offset)) = vbs.0[i] {
-                // TODO(fkaa): assign vertex buffers depending on what slots are
-                //             occupied? is it possible?
-                self.encoder.set_vertex_buffer((MTL_MAX_BUFFER_BINDINGS - 1) as u64 - vb_count, offset as u64, unsafe { *(buffer.0).0 });
-                vb_count += 1;
-            }
-        }
-    }
-
-    fn bind_constant_buffers(&mut self, cbs: &[pso::ConstantBufferParam<Resources>]) {
-        for &stage in [Stage::Vertex, Stage::Pixel].iter() {
-            let mask = stage.into();
-            for cb in cbs.iter() {
-                if cb.1.contains(mask) {
-                    match stage {
-                        Stage::Vertex => {
-                            self.encoder.set_vertex_buffer(cb.2 as u64, 0, unsafe { *((cb.0).0).0 });
-                        },
-                        Stage::Pixel => {
-                            self.encoder.set_fragment_buffer(cb.2 as u64, 0, unsafe { *((cb.0).0).0 });
-                        },
-                        _ => { unimplemented!() }
+            // FIXME: multiple buffers signaling!
+            let signal_block = if !submit.signal_semaphores.is_empty() {
+                let semaphores_copy: Vec<_> = submit.signal_semaphores.iter().map(|semaphore| {
+                    semaphore.0
+                }).collect();
+                Some(ConcreteBlock::new(move |cb: *mut ()| -> () {
+                    for semaphore in semaphores_copy.iter() {
+                        native::dispatch_semaphore_signal(*semaphore);
                     }
-                }
-            }
-        }
-    }
-
-    fn bind_global_constant(&mut self, _gc: shade::Location, _value: shade::UniformValue) {
-        unimplemented!()
-    }
-
-    fn bind_resource_views(&mut self, rvs: &[pso::ResourceViewParam<Resources>]) {
-        for &stage in [Stage::Vertex, Stage::Pixel].iter() {
-            let mask = stage.into();
-            for view in rvs.iter() {
-                if view.1.contains(mask) {
-                    match stage {
-                        Stage::Vertex => {
-                            self.encoder.set_vertex_texture(view.2 as u64, unsafe { *(view.0).0 });
-                        },
-                        Stage::Pixel => {
-                            self.encoder.set_fragment_texture(view.2 as u64, unsafe { *(view.0).0 });
-                        },
-                        _ => { unimplemented!() }
-                    }
-                }
-            }
-        }
-    }
-
-    fn bind_unordered_views(&mut self, _uvs: &[pso::UnorderedViewParam<Resources>]) {
-        // TODO: UAVs
-    }
-
-    fn bind_samplers(&mut self, ss: &[pso::SamplerParam<Resources>]) {
-        use std::f32;
-
-        for &stage in [Stage::Vertex, Stage::Pixel].iter() {
-            let mask = stage.into();
-            for sampler in ss.iter() {
-                if sampler.1.contains(mask) {
-                    match stage {
-                        Stage::Vertex => {
-                            self.encoder.set_vertex_sampler_state(sampler.2 as u64, (sampler.0).0);
-                        },
-                        Stage::Pixel => {
-                            self.encoder.set_fragment_sampler_state(sampler.2 as u64, (sampler.0).0)
-                        },
-                        _ => { unimplemented!() }
-                    }
-                }
-            }
-        }
-    }
-
-    fn bind_pixel_targets(&mut self, targets: pso::PixelTargetSet<Resources>) {
-        // TODO(fkaa): cache here to see if we're actually changing targets!
-        self.encoder.end_encoding();
-        self.should_restore = true;
-
-        let render_pass_descriptor = MTLRenderPassDescriptor::new();
-
-        for i in 0..MAX_COLOR_TARGETS {
-            if let Some(color) = targets.colors[i] {
-                let attachment = render_pass_descriptor.color_attachments().object_at(i);
-                attachment.set_texture(unsafe { *(color.0) });
-                attachment.set_store_action(MTLStoreAction::Store); // TODO: Multisample?
-
-                if let Some(clear_color) = self.rtv_clear.remove(&color) {
-                    // This attachement will handle the desired clear option
-                    attachment.set_load_action(MTLLoadAction::Clear);
-                    attachment.set_clear_color(clear_color);
-                } else {
-                    // if no clear has been specified, we simply load the
-                    // previous content
-                    attachment.set_load_action(MTLLoadAction::Load);
-                }
-            }
-        }
-
-        if let Some(depth) = targets.depth {
-            let stencil_attachment = if let Some(stencil) = targets.stencil {
-                debug_assert!(depth == stencil, "Depth and stencil targets must be the same if both are present");
-                Some(render_pass_descriptor.stencil_attachment())
+                }).copy())
             } else {
                 None
             };
 
-            let attachment = render_pass_descriptor.depth_attachment();
-            attachment.set_texture(unsafe { *(depth.0) });
-            if let Some(stencil_attachment) = stencil_attachment {
-                stencil_attachment.set_texture(unsafe { *(depth.0) });
-            }
-
-            if let Some(layer) = depth.1 {
-                attachment.set_slice(layer as u64);
-                if let Some(stencil_attachment) = stencil_attachment {
-                    stencil_attachment.set_slice(layer as u64);
+            for buffer in submit.cmd_buffers {
+                let command_buffer = buffer.get_info().command_buffer;
+                if let Some(ref signal_block) = signal_block {
+                    msg_send![command_buffer.0, addCompletedHandler: signal_block.deref() as *const _];
                 }
-            }
-
-            // do we need to handle any other cases?
-            attachment.set_store_action(MTLStoreAction::Store);
-            if let Some(stencil_attachment) = stencil_attachment {
-                stencil_attachment.set_store_action(MTLStoreAction::Store);
-            }
-
-            if let Entry::Occupied(mut clear_entry) = self.dsv_clear.entry(depth) {
-                if let Some(depth_value) = clear_entry.get().0 {
-                    attachment.set_load_action(MTLLoadAction::Clear);
-                    attachment.set_clear_depth(depth_value as f64);
-                } else {
-                     attachment.set_load_action(MTLLoadAction::Load);
-                }
-
-                // It may be the case that we have a stencil clear command for this buffer
-                // queued but a stencil attachment is not requested. In that case, the
-                // stencil clear must be preserved for future attachments, but the depth
-                // clear has been processed already and must be removed.
-                if let Some(stencil_value) = clear_entry.get().1 {
-                    if let Some(stencil_attachment) = stencil_attachment {
-                        stencil_attachment.set_load_action(MTLLoadAction::Clear);
-                        stencil_attachment.set_clear_stencil(stencil_value as u32);
-                        clear_entry.remove();
-                    } else {
-                        clear_entry.get_mut().0 = None;
+                // only append the fence handler to the last command buffer
+                if submit as *const _ == submit_infos.last().unwrap() as *const _ &&
+                   buffer as *const _ == submit.cmd_buffers.last().unwrap() as *const _ {
+                    if let Some(ref fence) = fence {
+                        let value_ptr = fence.0.clone();
+                        let fence_block = ConcreteBlock::new(move |cb: *mut ()| -> () {
+                            *value_ptr.lock().unwrap() = true;
+                        }).copy();
+                        msg_send![command_buffer.0, addCompletedHandler: fence_block.deref() as *const _];
                     }
-                } else {
-                    if let Some(stencil_attachment) = stencil_attachment {
-                        stencil_attachment.set_load_action(MTLLoadAction::Load);
-                    }
-                    clear_entry.remove();
                 }
-            } else {
-                // see above
-                attachment.set_load_action(MTLLoadAction::Load);
-                if let Some(stencil_attachment) = stencil_attachment {
-                    stencil_attachment.set_load_action(MTLLoadAction::Load);
-                }
+                command_buffer.commit();
             }
-        }
+        }*/
+    }
+}
 
-        self.encoder.set_render_pass_descriptor(render_pass_descriptor);
-        if let Some(dim) = targets.dimensions {
-            self.encoder.set_viewport(MTLViewport {
-                originX: 0f64,
-                originY: 0f64,
-                width: dim.0 as f64,
-                height: dim.1 as f64,
-                znear: 0f64,
-                zfar: 1f64
+impl core::RawCommandPool<Backend> for CommandPool {
+    fn reset(&mut self) {
+        self.active_buffers.clear();
+    }
+
+    unsafe fn from_queue(queue: &CommandQueue, flags: CommandPoolCreateFlags) -> Self {
+        unimplemented!()
+    }
+
+    fn allocate(&mut self, num: usize) -> Vec<CommandBuffer> {
+        unimplemented!()
+        /*unsafe {
+            // TODO: maybe use unretained command buffer for efficiency?
+            let command_buffer = self.queue.queue.new_command_buffer(); // Returns retained
+            defer_on_unwind! { command_buffer.release() }
+
+            self.active_buffers.push(CommandBuffer {
+                command_buffer,
+                encoder_state: EncoderState::None,
+                viewport: None,
+                scissors: None,
+                pipeline_state: None,
+                vertex_buffers: Vec::new(),
+                descriptor_sets: Vec::new(),
             });
+            Encoder::new(self.active_buffers.last_mut().unwrap())
+        }*/
+    }
+
+    /// Free command buffers which are allocated from this pool.
+    unsafe fn free(&mut self, buffers: Vec<CommandBuffer>) {
+        unimplemented!()
+    }
+}
+
+impl core::SubpassCommandPool<Backend> for CommandPool {
+}
+
+impl CommandBuffer {
+    #[inline]
+    fn inner(&mut self) -> &mut CommandBufferInner {
+        unsafe {
+            &mut *self.0.get()
         }
     }
 
-    fn bind_index(&mut self, buf: Buffer, idx_type: IndexType) {
-        use map::map_index_type;
-
-        // TODO(fkaa): pass wrapper instead
-        self.encoder.set_index_buffer(unsafe { *(buf.0).0 }, map_index_type(idx_type));
-    }
-
-    fn set_scissor(&mut self, rect: target::Rect) {
-        // TODO(fkaa): why are getting 1x1 scissor?
-        /*self.encoder.set_scissor_rect(MTLScissorRect {
-            x: rect.x as u64,
-            y: rect.y as u64,
-            width: (rect.x + rect.w) as u64,
-            height: (rect.y + rect.h) as u64,
-        });*/
-    }
-
-    fn set_ref_values(&mut self, vals: state::RefValues) {
-        // FIXME: wrong types?
-        self.encoder.set_stencil_front_back_reference_value(vals.stencil.0 as u32, vals.stencil.1 as u32);
-
-        // TODO: blend/stencil
-    }
-
-    #[allow(unused_variables)]
-    fn copy_buffer(&mut self, src: Buffer, dst: Buffer,
-                   src_offset_bytes: usize, dst_offset_bytes: usize,
-                   size_bytes: usize) {
-        unimplemented!()
-    }
-
-    #[allow(unused_variables)]
-    fn copy_buffer_to_texture(&mut self, src: Buffer, src_offset_bytes: usize,
-                              dst: Texture,
-                              kind: texture::Kind,
-                              face: Option<texture::CubeFace>,
-                              img: texture::RawImageInfo) {
-        unimplemented!()
-    }
-
-    #[allow(unused_variables)]
-    fn copy_texture_to_buffer(&mut self,
-                              src: Texture,
-                              kind: texture::Kind,
-                              face: Option<texture::CubeFace>,
-                              img: texture::RawImageInfo,
-                              dst: Buffer, dst_offset_bytes: usize) {
-        unimplemented!()
-    }
-
-    fn update_buffer(&mut self, buf: Buffer, data: &[u8], offset: usize) {
-        use map::{map_buffer_usage};
-
-        debug_assert!(!unsafe { *(*(buf.0).0) }.is_null(), "Buffer must be non-nil");
-
-        let b = unsafe { *(buf.0).0 };
-
-        // we can create a new buffer to avoid synchronization when:
-        //   * buffer is write-only
-        //   * we are replacing all the contents of the buffer
-        //
-        // TODO(fkaa): maybe have an upper limit to prevent from thrashing?
-        if offset == 0 && data.len() == b.length() as usize {
-            unsafe {
-                // TODO(fkaa): ensure the creation flags are identical
-                *(buf.0).0 = self.device.new_buffer_with_data(
-                    data.as_ptr() as _,
-                    data.len() as _,
-                    map_buffer_usage(buf.1, buf.2));
-
-                // invalidate old buffer in cache
-                self.encoder.invalidate_buffer(b);
+    fn encode_blit(&mut self) -> MTLBlitCommandEncoder {
+        unsafe {
+            match self.inner().encoder_state {
+                EncoderState::None => {},
+                EncoderState::Blit(blit_encoder) => return blit_encoder,
+                EncoderState::Render(render_encoder) => {
+                    render_encoder.end_encoding();
+                    render_encoder.release();
+                }
             }
+
+            let blit_encoder = self.inner().command_buffer.new_blit_command_encoder(); // Returns retained
+            self.inner().encoder_state = EncoderState::Blit(blit_encoder);
+            blit_encoder
+        }
+    }
+}
+
+impl core::RawCommandBuffer<Backend> for CommandBuffer {
+    fn begin(&mut self) {
+        unimplemented!()
+    }
+
+    fn finish(&mut self) {
+        match self.inner().encoder_state {
+            EncoderState::None => {},
+            EncoderState::Blit(blit_encoder) => blit_encoder.end_encoding(),
+            EncoderState::Render(render_encoder) => render_encoder.end_encoding(),
+        }
+    }
+
+    fn reset(&mut self, release_resources: bool) {
+        unimplemented!()
+    }
+
+    fn pipeline_barrier(
+        &mut self,
+        stages: Range<pso::PipelineStage>,
+        barriers: &[memory::Barrier<Backend>],
+    ) {
+        unimplemented!()
+    }
+
+    fn fill_buffer(
+        &mut self,
+        buffer: &native::Buffer,
+        range: Range<u64>,
+        data: u32,
+    ) {
+        unimplemented!()
+    }
+
+    fn update_buffer(
+        &mut self,
+        buffer: &native::Buffer,
+        offset: u64,
+        data: &[u8],
+    ) {
+        unimplemented!()
+    }
+
+    fn clear_color(
+        &mut self,
+        rtv: &native::RenderTargetView,
+        layout: ImageLayout,
+        color: ClearColor,
+    ) {
+        unimplemented!()
+    }
+
+    fn clear_depth_stencil(
+        &mut self,
+        dsv: &native::DepthStencilView,
+        layout: ImageLayout,
+        depth_value: Option<target::Depth>,
+        stencil_value: Option<target::Stencil>,
+    ) {
+        unimplemented!()
+    }
+
+    fn resolve_image(
+        &mut self,
+        src: &native::Image,
+        src_layout: ImageLayout,
+        dst: &native::Image,
+        dst_layout: ImageLayout,
+        regions: &[ImageResolve],
+    ) {
+        unimplemented!()
+    }
+
+    fn bind_index_buffer(&mut self, view: IndexBufferView<Backend>) {
+        unimplemented!()
+    }
+
+    fn bind_vertex_buffers(&mut self, buffer: pso::VertexBufferSet<Backend>) {
+        self.inner().vertex_buffers.clear();
+        self.inner().vertex_buffers.extend(buffer.0.iter().map(|&(buffer, offset)| (buffer.0, offset)));
+    }
+
+    fn set_viewports(&mut self, rects: &[Viewport]) {
+        if rects.len() != 1 {
+            panic!("Metal supports only one viewport");
+        }
+        let rect = &rects[0];
+        self.inner().viewport = Some(MTLViewport {
+            originX: rect.x as f64,
+            originY: rect.y as f64,
+            width: rect.w as f64,
+            height: rect.h as f64,
+            znear: 0.0,
+            zfar: 1.0,
+        });
+    }
+
+    fn set_scissors(&mut self, rects: &[target::Rect]) {
+        if rects.len() != 1 {
+            panic!("Metal supports only one scissor");
+        }
+        let rect = &rects[0];
+        self.inner().scissors = Some(MTLScissorRect {
+            x: rect.x as NSUInteger,
+            y: rect.y as NSUInteger,
+            width: rect.w as NSUInteger,
+            height: rect.h as NSUInteger,
+        });
+    }
+
+    fn set_stencil_reference(&mut self, front: target::Stencil, back: target::Stencil) {
+        unimplemented!()
+    }
+
+    fn set_blend_constants(&mut self, color: target::ColorValue) {
+        unimplemented!()
+    }
+
+    fn begin_renderpass(
+        &mut self,
+        render_pass: &native::RenderPass,
+        frame_buffer: &native::FrameBuffer,
+        render_area: target::Rect,
+        clear_values: &[ClearValue],
+        first_subpass: SubpassContents,
+    ) {
+        unimplemented!()
+    }
+    
+    fn next_subpass(&mut self, contents: SubpassContents) {
+        unimplemented!()
+    }
+    
+    fn end_renderpass(&mut self) {
+        unimplemented!()
+    }
+
+    fn bind_graphics_pipeline(&mut self, pipeline: &native::GraphicsPipeline) {
+        self.inner().pipeline_state = Some(pipeline.0);
+    }
+
+    fn bind_graphics_descriptor_sets(
+        &mut self,
+        layout: &native::PipelineLayout,
+        first_set: usize,
+        sets: &[&native::DescriptorSet],
+    ) {
+        while self.inner().descriptor_sets.len() < first_set + sets.len() {
+            self.inner().descriptor_sets.push(None);
+        }
+        for (out, &set) in self.inner().descriptor_sets[first_set ..].iter_mut().zip(sets) {
+            *out = Some(set.clone());
+        }
+    }
+
+    fn bind_compute_pipeline(&mut self, pipeline: &native::ComputePipeline) {
+        unimplemented!()
+    }
+    
+    fn dispatch(&mut self, x: u32, y: u32, z: u32) {
+        unimplemented!()
+    }
+    
+    fn dispatch_indirect(&mut self, buffer: &native::Buffer, offset: u64) {
+        unimplemented!()
+    }
+    
+    fn copy_buffer(
+        &mut self,
+        src: &native::Buffer,
+        dst: &native::Buffer,
+        regions: &[BufferCopy],
+    ) {
+        unimplemented!()
+    }
+    
+    fn copy_image(
+        &mut self,
+        src: &native::Image,
+        src_layout: ImageLayout,
+        dst: &native::Image,
+        dst_layout: ImageLayout,
+        regions: &[ImageCopy],
+    ) {
+        unimplemented!()
+    }
+    
+    fn copy_buffer_to_image(
+        &mut self,
+        src: &native::Buffer,
+        dst: &native::Image,
+        dst_layout: ImageLayout,
+        regions: &[BufferImageCopy],
+    ) {
+        let encoder = self.encode_blit();
+        let extent = unsafe { MTLSize {
+            width: dst.0.width(),
+            height: dst.0.height(),
+            depth: dst.0.depth(),
+        }};
+        // FIXME: layout
+
+        for region in regions {
+            let image_offset = &region.image_offset;
+
+            // TODO multiple layers
+            assert!((region.image_subresource.1).len() == 1, "multiple layer copies not implemented");
+            unsafe {
+                msg_send![encoder.0,
+                    copyFromBuffer: (src.0).0
+                    sourceOffset: region.buffer_offset as NSUInteger
+                    sourceBytesPerRow: region.buffer_row_pitch as NSUInteger
+                    sourceBytesPerImage: region.buffer_slice_pitch as NSUInteger
+                    sourceSize: extent
+                    toTexture: (dst.0).0
+                    destinationSlice: (region.image_subresource.1).start as NSUInteger // TODO: not sure this is right
+                    destinationLevel: region.image_subresource.0 as NSUInteger
+                    destinationOrigin: MTLOrigin { x: image_offset.x as NSUInteger, y: image_offset.y as NSUInteger, z: image_offset.z as NSUInteger }
+                ]
+            }
+        }
+    }
+    
+    fn copy_image_to_buffer(
+        &mut self,
+        src: &native::Image,
+        src_layout: ImageLayout,
+        dst: &native::Buffer,
+        regions: &[BufferImageCopy],
+    ) {
+        unimplemented!()
+    }
+    
+    fn draw(
+        &mut self,
+        vertices: Range<VertexCount>,
+        instances: Range<InstanceCount>,
+    ) {
+        unimplemented!()
+    }
+    
+    fn draw_indexed(
+        &mut self,
+        indeces: Range<IndexCount>,
+        base_vertex: VertexOffset,
+        instances: Range<InstanceCount>,
+    ) {
+        unimplemented!()
+    }
+    
+    fn draw_indirect(
+        &mut self,
+        buffer: &native::Buffer,
+        offset: u64,
+        draw_count: u32,
+        stride: u32,
+    ) {
+        unimplemented!()
+    }
+    
+    fn draw_indexed_indirect(
+        &mut self,
+        buffer: &native::Buffer,
+        offset: u64,
+        draw_count: u32,
+        stride: u32,
+    ) {
+        unimplemented!()
+    }
+}
+
+/*
+pub struct RenderPassInlineBuffer {
+    render_encoder: MTLRenderCommandEncoder,
+}
+
+impl Drop for RenderPassInlineBuffer {
+    fn drop(&mut self) {
+        unsafe { self.render_encoder.release(); }
+    }
+}
+
+impl RenderPassInlineBuffer {
+    #[cfg(feature = "argument_buffer")]
+    fn bind_descriptor_set(encoder: MTLRenderCommandEncoder, slot: u64, set: &native::DescriptorSet) {
+        if set.stage_flags.contains(shade::STAGE_VERTEX) {
+            encoder.set_vertex_buffer(slot, set.offset, set.buffer)
+        }
+        if set.stage_flags.contains(shade::STAGE_PIXEL) {
+            encoder.set_fragment_buffer(slot, set.offset, set.buffer)
+        }
+    }
+
+    #[cfg(not(feature = "argument_buffer"))]
+    fn bind_descriptor_set(encoder: MTLRenderCommandEncoder, _slot: u64, set: &native::DescriptorSet) {
+        use native::DescriptorSetBinding::*;
+        let set = set.inner.lock().unwrap();
+
+        for (&binding, values) in set.bindings.iter() {
+            let layout = set.layout.iter().find(|x| x.binding == binding).unwrap();
+
+            if layout.stage_flags.contains(pso::STAGE_FRAGMENT) {
+                match *values {
+                    Sampler(ref samplers) => {
+                        if samplers.len() > 1 {
+                            unimplemented!()
+                        }
+
+                        let sampler = samplers[0];
+                        encoder.set_fragment_sampler_state(binding as u64, sampler);
+                    },
+                    SampledImage(ref images) => {
+                        if images.len() > 1 {
+                            unimplemented!()
+                        }
+
+                        let (image, layout) = images[0]; // TODO: layout?
+                        encoder.set_fragment_texture(binding as u64, image);
+                    },
+                    _ => unimplemented!(),
+                }
+            }
+            if layout.stage_flags.contains(pso::STAGE_VERTEX) {
+                unimplemented!()
+            }
+        }
+    }
+
+    fn begin_renderpass(&mut self,
+                        render_pass: &native::RenderPass,
+                        frame_buffer: &native::FrameBuffer,
+                        render_area: target::Rect,
+                        clear_values: &[ClearValue],
+                        first_subpass: command::SubpassContents,
+    ) {
+        unsafe {
+            // FIXME: subpasses
+
+            let pass_descriptor = frame_buffer.0;
+            // TODO: we may want to copy here because we will modify the Framebuffer (by setting
+            // clear colors)
+            // TODO: validate number of clear colors
+            for (i, value) in clear_values.iter().enumerate() {
+                let color_desc = pass_descriptor.color_attachments().object_at(i);
+                let mtl_color = match *value {
+                    ClearValue::Color(ClearColor::Float(values)) => MTLClearColor::new(
+                        values[0] as f64,
+                        values[1] as f64,
+                        values[2] as f64,
+                        values[3] as f64,
+                    ),
+                    _ => unimplemented!(),
+                };
+                color_desc.set_clear_color(mtl_color);
+            }
+
+            let render_encoder = self.command_buffer.new_render_command_encoder(pass_descriptor);
+            defer_on_unwind! { render_encoder.release() };
+
+            // Apply previously bound values for this command buffer
+            if let Some(viewport) = self.viewport {
+                render_encoder.set_viewport(viewport);
+            }
+            if let Some(scissors) = self.scissors {
+                render_encoder.set_scissor_rect(scissors);
+            }
+            if let Some(pipeline_state) = self.pipeline_state {
+                render_encoder.set_render_pipeline_state(pipeline_state);
+            } else {
+                panic!("missing bound pipeline state object");
+            }
+            for (i, &(buffer, offset)) in self.vertex_buffers.iter().enumerate() {
+                render_encoder.set_vertex_buffer(i as u64, offset as u64, buffer);
+            }
+            // Interpret descriptor sets
+            for (i, set_maybe) in self.descriptor_sets.iter().enumerate() {
+                if let &Some(ref set) = set_maybe {
+                    Self::bind_descriptor_set(render_encoder, i as u64, set);
+                }
+            }
+
+            unimplemented!()
+            /*RenderPassInlineBuffer {
+                render_encoder,
+            }*/
+        }
+    }
+
+    /* renderpass: fn finish(&mut self,
+              command_buffer: &mut CommandBuffer,
+              render_pass: &native::RenderPass,
+              framebuffer: &native::FrameBuffer) {
+        unsafe {
+            self.render_encoder.end_encoding();
+        }
+    }*/
+
+    fn draw(
+        encoder: &mut self,
+        vertices: Range<VertexCount>,
+        instances: Range<InstanceCount>,
+    ) {
+        if let Some(instance) = instance {
+            unimplemented!()
         } else {
-            // TODO(fkaa): how do we (eventually) handle updating buffers when
-            //             we have multiple encoders? enqueue cmd buffers and
-            //             rely on their synchronization?
-
-            // TODO(fkaa): perhaps we can keep track of in-use buffers in some
-            //             higher-level place and query here to prevent from
-            //             stalling encoding when buffer is not in use
-
-            // TODO(fkaa): need to have a better way of keeping track of
-            //             internal cmd buffers, not good to commit/request
-            //             constantly
-
-            // FIXME: slow :-(
-            self.encoder.end_encoding();
-            self.encoder.commit_command_buffer(CAMetalDrawable::nil(), true);
-            self.should_restore = true;
-
-            let contents = b.contents();
-
-            unsafe {
-                let dst = (contents as *mut u8).offset(offset as isize);
-                ptr::copy(data.as_ptr(), dst, data.len());
-
-                // TODO(fkaa): notify *only if* buffer has managed storage mode:
-                // b.did_modify_range(NSRange::new(offset as u64, data.len() as u64));
-            }
-        }
-    }
-
-    fn update_texture(&mut self,
-                      tex: Texture,
-                      kind: texture::Kind,
-                      face: Option<texture::CubeFace>,
-                      data: &[u8],
-                      info: texture::RawImageInfo) {
-        unimplemented!()
-    }
-
-    fn generate_mipmap(&mut self, _srv: Srv) {
-        unimplemented!()
-    }
-
-    fn clear_color(&mut self, target: Rtv, value: command::ClearColor) {
-        let double_value = match value {
-            command::ClearColor::Float(val) => MTLClearColor::new(val[0] as f64, val[1] as f64, val[2] as f64, val[3] as f64),
-            command::ClearColor::Int(val) => MTLClearColor::new(val[0] as f64, val[1] as f64, val[2] as f64, val[3] as f64),
-            command::ClearColor::Uint(val) => MTLClearColor::new(val[0] as f64, val[1] as f64, val[2] as f64, val[3] as f64),
-        };
-
-        self.rtv_clear.insert(target, double_value);
-    }
-
-    fn clear_depth_stencil(&mut self, target: Dsv,
-                           depth: Option<target::Depth>, stencil: Option<target::Stencil>) {
-        match self.dsv_clear.entry(target) {
-            Entry::Occupied(mut entry) => {
-                let new_depth = depth.or(entry.get().0);
-                let new_stencil = stencil.or(entry.get().1);
-                entry.insert((new_depth, new_stencil));
-            },
-            Entry::Vacant(entry) => { entry.insert((depth, stencil)); },
-        }
-    }
-
-    fn call_draw(&mut self, start: VertexCount, count: VertexCount, instances: Option<command::InstanceParams>) {
-        self.ensure_render_encoder();
-
-        match instances {
-            Some((ninst, offset)) => {
-                self.encoder.draw_instanced(count as u64,
-                                            ninst as u64,
-                                            start as u64);
-            },
-            None => {
-                self.encoder.draw(start as u64,
-                                  count as u64);
-            }
-        }
-    }
-
-    fn call_draw_indexed(&mut self, start: VertexCount, count: VertexCount,
-                         base: VertexCount, instances: Option<command::InstanceParams>) {
-        self.ensure_render_encoder();
-
-        match instances {
-            Some((ninst, offset)) => {
-                self.encoder.draw_indexed_instanced(count as u64,
-                                                    start as u64,
-                                                    ninst as u64,
-                                                    base as i64,
-                                                    offset as u64);
-            },
-            None => {
-                self.encoder.draw_indexed(count as u64,
-                                          start as u64);
-            }
+            // FIXME: primitive type
+            encoder.pass_buffer.render_encoder.draw_primitives(MTLPrimitiveType::Triangle, start as u64, count as u64);
         }
     }
 }
-
-pub struct SubpassCommandBuffer;
-
-impl core::CommandBuffer<Backend> for SubpassCommandBuffer {
-    unsafe fn end(&mut self) -> SubmitInfo {
-        unimplemented!()
-    }
-}
+*/
